pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "g1-hybrid-prior"
version = "0.1.0"
description = "Hybrid prior from G1 humanoid logs"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

authors = [
  { name = "Valerio Belli", email = "valeriobelli26@gmail.com" }
]

dependencies = [
  "numpy>=1.23,<2", 
  "pyyaml>=6.0",
  "ipykernel>=7.1.0",
  "ipython>=8.37.0",
  "vector-quantize-pytorch>=1.27.19",
]

[project.optional-dependencies]
dev = [
  "black",
  "ruff",
  "pytest",
]

[project.scripts]
g1-hybrid-prior = "g1_hybrid_prior.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']
